fix compile errors and warnings for linux
use system glx.h

use -Wno-unused-result

fix pointer and int sizes

+++ paintball2/client/cl_scrn.c
@@ -901,17 +901,17 @@ void SCR_Loading_f (void)
 SCR_TimeRefresh_f
 ================
 */
-int entitycmpfnc (const entity_t *a, const entity_t *b)
+off_t entitycmpfnc (const entity_t *a, const entity_t *b)
 {
 	// all other models are sorted by model then skin
 	if (a->model == b->model)
 	{
 		 // todo - for multi-skin meshes, this probably isn't going to buy us anything...
-		return ((int)a->skins[0] - (int)b->skins[0]);
+		return ((intptr_t)a->skins[0] - (intptr_t)b->skins[0]);
 	}
 	else
 	{
-		return ((int)a->model - (int)b->model);
+		return ((intptr_t)a->model - (intptr_t)b->model);
 	}
 }
